requirements.txt
fastapi==0.75.1
pandas==1.4.2
sqlalchemy==1.4.35
catboost==1.0.6
numpy==1.22.4
pydantic==1.9.1
scikit_learn==1.1.1
psycopg2-binary==2.9.3
uvicorn==0.16.0
python-dotenv==0.20.0
pytorch==2.1.2+cu118
optuna==3.3.0
nltk==3.7
transformers==4.24.0